server/requirements.txt
aiofiles==22.1.0
alembic==1.4.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
bcrypt==4.0.1
cffi==1.15.1
click==7.1.2
fastapi==0.63.0
h11==0.11.0
Jinja2==2.11.3
Mako==1.1.3
MarkupSafe==1.1.1
passlib==1.7.4
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
pydantic==1.7.3
PyJWT==2.6.0
python-dateutil==2.8.1
python-editor==1.0.4
python-multipart==0.0.5
six==1.15.0
SQLAlchemy==1.3.22
starlette==0.13.6
uvicorn==0.13.2
scikit-learn==1.2.0
